challenge 10 about do the currency from reais to dollars

+++ problems/4th class/challenge10.py
@@ -1,2 +1,2 @@
-x = int(input('How much money do you have R$: '))
+x = float(input('How much money do you have R$: '))
 print('You can buy {:.2f} dollars'.format(x / 3.27))
